pyproject.toml
[project]
name = "ping-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "async-pymongo>=0.1.11",
    "fastapi[standard]>=0.115.12",
    "neo4j>=5.28.1",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "python-jose>=3.4.0",
]
